# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Full Stack FastAPI Project", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/clients/": {
      "get": {
        "tags": ["clients"],
        "summary": "Get All Clients",
        "description": "Get all clients.",
        "operationId": "clients-get_all_clients",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/Client" },
                  "type": "array",
                  "title": "Response Clients-Get All Clients"
                }
              }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      },
      "post": {
        "tags": ["clients"],
        "summary": "Create Client",
        "description": "Create new client.",
        "operationId": "clients-create_client",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ClientCreate" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Client" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      }
    },
    "/api/v1/clients/{client_id}": {
      "put": {
        "tags": ["clients"],
        "summary": "Update Client",
        "description": "Update an existing client.",
        "operationId": "clients-update_client",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Client Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ClientUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Client" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["clients"],
        "summary": "Get Client",
        "description": "Get a specific client by client_id.",
        "operationId": "clients-get_client",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Client Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Client" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["clients"],
        "summary": "Delete Client",
        "description": "Delete a client by client_id.",
        "operationId": "clients-delete_client",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Client Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Client" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/": {
      "get": {
        "tags": ["appointments"],
        "summary": "Get All Appointments",
        "description": "Get all appointments.",
        "operationId": "appointments-get_all_appointments",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/Appointment" },
                  "type": "array",
                  "title": "Response Appointments-Get All Appointments"
                }
              }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      },
      "post": {
        "tags": ["appointments"],
        "summary": "Create Appointment",
        "description": "Create a new appointment.",
        "operationId": "appointments-create_appointment",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AppointmentCreate" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Appointment" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      }
    },
    "/api/v1/appointments/{appointment_id}": {
      "put": {
        "tags": ["appointments"],
        "summary": "Update Appointment",
        "description": "Update an existing appointment.",
        "operationId": "appointments-update_appointment",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Appointment Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AppointmentUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Appointment" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["appointments"],
        "summary": "Get Appointment",
        "description": "Get a specific appointment by appointment_id.",
        "operationId": "appointments-get_appointment",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Appointment Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Appointment" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["appointments"],
        "summary": "Delete Appointment",
        "description": "Delete an appointment by appointment_id.",
        "operationId": "appointments-delete_appointment",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Appointment Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Appointment" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Appointment": {
        "properties": {
          "appointment_date": {
            "type": "string",
            "format": "date-time",
            "title": "Appointment Date"
          },
          "price": { "type": "number", "title": "Price" },
          "status": { "type": "string", "maxLength": 50, "title": "Status" },
          "client_notes": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Client Notes"
          },
          "internal_notes": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Internal Notes"
          },
          "duration": { "type": "integer", "title": "Duration" },
          "client_id": { "type": "string", "format": "uuid", "title": "Client Id" },
          "provider_id": { "type": "string", "format": "uuid", "title": "Provider Id" },
          "service_id": { "type": "string", "format": "uuid", "title": "Service Id" },
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["price", "status", "duration", "client_id", "provider_id", "service_id"],
        "title": "Appointment"
      },
      "AppointmentCreate": {
        "properties": {
          "appointment_date": {
            "type": "string",
            "format": "date-time",
            "title": "Appointment Date"
          },
          "price": { "type": "number", "title": "Price" },
          "status": { "type": "string", "maxLength": 50, "title": "Status" },
          "client_notes": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Client Notes"
          },
          "internal_notes": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Internal Notes"
          },
          "duration": { "type": "integer", "title": "Duration" },
          "client_id": { "type": "string", "format": "uuid", "title": "Client Id" },
          "provider_id": { "type": "string", "format": "uuid", "title": "Provider Id" },
          "service_id": { "type": "string", "format": "uuid", "title": "Service Id" }
        },
        "type": "object",
        "required": ["price", "status", "duration", "client_id", "provider_id", "service_id"],
        "title": "AppointmentCreate"
      },
      "AppointmentUpdate": {
        "properties": {
          "appointment_date": {
            "type": "string",
            "format": "date-time",
            "title": "Appointment Date"
          },
          "price": { "type": "number", "title": "Price" },
          "status": { "type": "string", "title": "Status" },
          "client_notes": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Client Notes"
          },
          "internal_notes": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Internal Notes"
          },
          "duration": { "type": "integer", "title": "Duration" },
          "client_id": { "type": "string", "format": "uuid", "title": "Client Id" },
          "provider_id": { "type": "string", "format": "uuid", "title": "Provider Id" },
          "service_id": { "type": "string", "format": "uuid", "title": "Service Id" }
        },
        "type": "object",
        "required": [
          "appointment_date",
          "price",
          "status",
          "duration",
          "client_id",
          "provider_id",
          "service_id"
        ],
        "title": "AppointmentUpdate"
      },
...
      "Client": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Last Name" },
          "email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
          "phone": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Phone" },
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["first_name", "last_name", "email", "phone"],
        "title": "Client"
      },
      "ClientCreate": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Last Name" },
          "email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
          "phone": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Phone" }
        },
        "type": "object",
        "required": ["first_name", "last_name", "email", "phone"],
        "title": "ClientCreate"
      },
      "ClientUpdate": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Last Name" },
          "email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
          "phone": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Phone" }
        },
        "type": "object",
        "required": ["first_name", "last_name", "email", "phone"],
        "title": "ClientUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": { "SupabaseBearer": { "type": "http", "scheme": "bearer" } }
  }
}